enhancement/updates (#46)
* started updating

* updated

* updates

* attestation

* fixed

* attestation

* removed

* updated

* fixed

* fix

* reformat

+++ Sources/BeaconChain/DataStructures/Operations/ProposerSlashing.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct ProposerSlashing {
-    let proposerIndex: UInt32 // @todo should be UInt24
+    let proposerIndex: UInt64
     let proposalData1: ProposalSignedData
     let proposalSignature1: Data
     let proposalData2: ProposalSignedData
