+++ Sources/BeaconChain/Types.swift
@@ -3,7 +3,7 @@ import Foundation
 typealias SlotNumber = UInt64
 typealias EpochNumber = UInt64
 typealias ShardNumber = UInt64
-typealias ValidatorIndex = UInt32 // @todo should be UInt24
+typealias ValidatorIndex = UInt64
 typealias Gwei = UInt64
 typealias Bytes32 = Data // @todo needs to be 32 fixed length data
 typealias BLSPubkey = Data // @todo needs to be 48 fixed length data
